// webapp/src/api/model/upgradeRecipe.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { UpgradeRecipeMaterialsItem } from './upgradeRecipeMaterialsItem';
import type { Upgrade } from './upgrade';

export interface UpgradeRecipe {
  id?: string;
  materials: UpgradeRecipeMaterialsItem[];
  requiredUpgrades: Upgrade[];
  requirement: string;
  upgrade: Upgrade;
}